requirements.txt
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
pydantic
pdfplumber
pytesseract
Pillow
openai>=1.32.0,<2.0.0
langchain_groq

pinecone-notebooks
langchain-openai
langchain-pinecone
sentence-transformers==2.2.2
chroma-hnswlib
InstructorEmbedding
jq
chromadb
